requirements.txt
bcrypt==4.0.1
certifi==2022.9.24
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-JWT-Extended==4.4.4
flask-serialize==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
permissive-dict==1.0.4
PyJWT==2.6.0
pymongo==4.3.2
Werkzeug==2.2.2

numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
